#227 moved all dom modules and utils to packages
