+++ packages/react-look/modules/core/renderer.js
@@ -1,8 +1,9 @@
-import StyleContainer from '../../api/dom/StyleContainer'
-import sortPseudoClasses from '../../utils/sortPseudoClasses'
-import isMediaQuery from '../../utils/isMediaQuery'
-import isPseudo from '../../utils/isPseudo'
 import _ from 'lodash'
+
+import StyleContainer from '../api/StyleContainer'
+import sortPseudoClasses from '../utils/sortPseudoClasses'
+import isMediaQuery from '../utils/isMediaQuery'
+import isPseudo from '../utils/isPseudo'
 
 /**
  * Extracts all possible dynamic styles out of a style object
