+++ packages/react-look/modules/utils/generateHashCode.js
