+++ packages/react-look/modules/utils/isPseudo.js
